+++ code/filesystem/fscore/fsc_shader.c
@@ -52,7 +52,7 @@ static int FSC_IndexShaderFileData( fsc_filesystem_t *fs, fsc_stackptr_t source_
 
 		while ( 1 ) {
 			// Load next token
-			shader_start_position = current_position - data;
+			shader_start_position = (unsigned int)( current_position - data );
 			FSC_ParseExt( token, &current_position, fsc_true );
 			if ( !*token ) {
 				// We reached the end of the shader file.
@@ -110,7 +110,7 @@ static int FSC_IndexShaderFileData( fsc_filesystem_t *fs, fsc_stackptr_t source_
 		new_shader->shader_name_ptr = FSC_StringRepositoryGetString( shader_name, &fs->string_repository );
 		new_shader->source_file_ptr = source_file_ptr;
 		new_shader->start_position = shader_start_position;
-		new_shader->end_position = current_position - data;
+		new_shader->end_position = (unsigned int)( current_position - data );
 
 		// Add shader to hash table
 		FSC_HashtableInsert( new_shader_ptr, hash, &fs->shaders );
